include calldata in generator plugins
