+++ src/groth16.js
@@ -20,4 +20,3 @@
 export {default as fullProve} from "./groth16_fullprove.js";
 export {default as prove} from "./groth16_prove.js";
 export {default as verify} from "./groth16_verify.js";
-export {default as exportSolidityCallData} from "./groth16_exportsoliditycalldata.js";
