+++ src/plonk.js
@@ -22,4 +22,3 @@ export {default as setup} from "./plonk_setup.js";
 export {default as fullProve} from "./plonk_fullprove.js";
 export {default as prove} from "./plonk_prove.js";
 export {default as verify} from "./plonk_verify.js";
-export {default as exportSolidityCallData} from "./plonk_exportsoliditycalldata.js";
