+++ src/zkey.js
@@ -29,3 +29,4 @@ export {default as exportJson} from "./zkey_export_json.js";
 export {default as bellmanContribute} from "./zkey_bellman_contribute.js";
 export {default as exportVerificationKey} from "./zkey_export_verificationkey.js";
 export {default as exportVerifier} from "./zkey_export_verifier.js";
+export {default as exportCalldata} from "./zkey_export_calldata.js";
