+++ src/zkey_export_calldata.js
@@ -1,0 +1,3 @@
+export default async function exportCalldata(proof, pub, funcs, logger) {
+    return funcs[proof.protocol](proof, pub, logger);
+}
